Calculation of the geodesic of a point to multiple points

This function calculates the geodesic of a point to multiple points given the coordinate information. It is a wrapper of geodist::geodist_vec().

dist_point_to_vector(plon, plat, vlon, vlat)

Arguments

plon

Numeric. The longitude of a point.

plat

Numeric. The latitude of a point.

vlon

Numeric. A vector of longitude values.

vlat

Numeric. A vector of latitude values.

Value

Numeric. The geodesic of a point to multiple points in meters.

Examples

# Define vlon and vlat
vlon <- c(141.12, 141.13)
vlat <- c(-37.1, -37.0)

# Calculate the geodesic
dist_point_to_vector(141.12, -37.1, vlon, vlat)
#> [1]     0.00 11148.63