Summarizing spatiotemporal clustering results

This function takes a spotoroo object to produce a summary of the clustering results. It can be called by summary.spotoroo().

summary_spotoroo(result, cluster = "all")

Arguments

result

spotoroo object. A result of a call to hotspot_cluster().

cluster

Character/Integer. If "all", summarize all clusters. If an integer vector is given, summarize corresponding clusters.

Value

No return value, called for side effects

Examples

# \donttest{

  # Time consuming functions (>5 seconds)

  # Get clustering results
  result <- hotspot_cluster(hotspots,
                           lon = "lon",
                           lat = "lat",
                           obsTime = "obsTime",
                           activeTime = 24,
                           adjDist = 3000,
                           minPts = 4,
                           minTime = 3,
                           ignitionCenter = "mean",
                           timeUnit = "h",
                           timeStep = 1)
#> 
#> ──────────────────────────────── SPOTOROO 0.1.2 ────────────────────────────────
#> 
#> ── Calling Core Function : `hotspot_cluster()` ──
#> 
#> ── "1" time index = 1 hour 
#> ✔ Transform observed time → time indexes
#> ℹ 970 time indexes found
#> 
#> ── activeTime = 24 time indexes | adjDist = 3000 meters 
#> ✔ Cluster
#> ℹ 16 clusters found (including noise)
#> 
#> ── minPts = 4 hot spots | minTime = 3 time indexes 
#> ✔ Handle noise
#> ℹ 6 clusters left
#> ℹ noise proportion : 0.935 %
#> 
#> ── ignitionCenter = "mean" 
#> ✔ Compute ignition points for clusters
#> ℹ average hot spots : 176.7
#> ℹ average duration : 131.9 hours
#> 
#> ── Time taken = 0 mins 1 sec for 1070 hot spots 
#> ℹ 0.001 secs per hot spot
#> 
#> ────────────────────────────────────────────────────────────────────────────────

  # Make a summary of all clusters
  summary_spotoroo(result)
#> 
#> ──────────────────────────────── SPOTOROO 0.1.2 ────────────────────────────────
#> 
#> ── Calling Core Function : `summary_spotoroo()` ──
#> 
#> CLUSTERS: ALL
#> OBSERVATIONS: 1070
#> FROM: 2019-12-29 13:10:00
#> TO:   2020-02-07 22:50:00
#> 
#> 
#> ── Clusters 
#> ℹ Number of clusters: 6
#> 
#> Observations in cluster
#>         Min.     1st Qu.        Mean     3rd Qu.        Max.
#>        111.0       131.0       176.7       233.2       256.0
#> Duration of cluster (hours)
#>         Min.     1st Qu.        Mean     3rd Qu.        Max.
#>        111.2       118.2       131.9       146.1       148.3
#> 
#> ── Hot spots (excluding noise) 
#> ℹ Number of hot spots: 1060
#> 
#> Distance to ignition points (m)
#>         Min.     1st Qu.        Mean     3rd Qu.        Max.
#>          0.0      2840.3      5058.2      6981.6     13452.7
#> Time from ignition (hours)
#>         Min.     1st Qu.        Mean     3rd Qu.        Max.
#>          0.0        25.2        62.5        98.2       148.3
#> 
#> ── Noise 
#> ℹ Number of noise points: 10 (0.93 %)
#> 
#> 
#> ────────────────────────────────────────────────────────────────────────────────

  # Make a summary of cluster 1 to 3
  summary_spotoroo(result, 1:3)
#> 
#> ──────────────────────────────── SPOTOROO 0.1.2 ────────────────────────────────
#> 
#> ── Calling Core Function : `summary_spotoroo()` ──
#> 
#> CLUSTERS: 1 2 3
#> OBSERVATIONS: 447
#> FROM: 2019-12-29 13:10:00
#> TO:   2020-02-01 05:20:00
#> 
#> 
#> ── Clusters 
#> ℹ Number of clusters: 3
#> 
#> Observations in cluster
#>         Min.     1st Qu.        Mean     3rd Qu.        Max.
#>        126.0       136.0       145.7       155.5       165.0
#> Duration of cluster (hours)
#>         Min.     1st Qu.        Mean     3rd Qu.        Max.
#>        116.2       131.2       136.9       147.3       148.3
#> 
#> ── Hot spots (excluding noise) 
#> ℹ Number of hot spots: 437
#> 
#> Distance to ignition points (m)
#>         Min.     1st Qu.        Mean     3rd Qu.        Max.
#>          0.0      2222.7      4788.2      6944.3     13452.7
#> Time from ignition (hours)
#>         Min.     1st Qu.        Mean     3rd Qu.        Max.
#>          0.0        14.0        61.3        97.5       148.3
#> 
#> ────────────────────────────────────────────────────────────────────────────────
# }