Calculation of the fire movement

get_fire_mov() calculates the movement of a single fire per step time indexes. It collects hotspots per step time indexes, then takes the mean or median of the longitude and latitude as the centre of the fire.

get_fire_mov(result, cluster, step = 1, method = "mean")

Arguments

result
spotoroo object; a result of a call to hotspot_cluster().

cluster
integer; the membership label of the cluster.

step
integer (>=0); step size used in the calculation of the fire movement.

method
character; "mean" or "median", method of the calculation of the centre of the fire.

Value

data.frame; the fire movement

membership : membership labels.

lon : longitude of the centre of the fire.

lat : latitude of the centre of the fire.

timeID : time indexes.

obsTime : observed time (approximated).

ignition : whether or not it is a ignition point.

Examples

result <- hotspot_cluster(hotspots_fin,
                          lon = "lon",
                          lat = "lat",
                          obsTime = "obsTime",
                          activeTime = 24,
                          adjDist = 3000,
                          minPts = 4,
                          minTime = 3,
                          ignitionCenter = "mean",
                          timeUnit = "h",
                          timeStep = 1)

#> 

#> ------------------------------ SPOTOROO 0.0.0.9000 -----------------------------

#> 

#> -- Calling Core Function : `hotspot_cluster()` --

#> 

#> -- 1 time index = 1 hours 

#> √ Transform observed time > time indexes

#> i 970 time indexes found

#> 

#> -- activeTime = 24 time indexes | adjDist = 3000 meters 

#> √ Cluster

#> i 16 clusters found (including noise)

#> 

#> -- minPts = 4 hotspots | minTime = 3 time indexes 

#> √ Handle noise

#> i 6 clusters left

#> i noise proportion : 0.935 % 

#> 

#> -- ignitionCenter = 'mean' 

#> √ Compute ignition points for clusters

#> i average hotspots : 176.7

#> i average duration : 131.9 hours

#> 

#> -- Time taken = 0 mins 8 secs for 1070 hotspots 

#> i 0.008 secs per hotspot

#> 

#> --------------------------------------------------------------------------------

get_fire_mov(result, cluster = 1, step = 1, method = "mean")

#>    membership      lon       lat timeID             obsTime ignition
#> 1           1 149.3000 -37.77000      1 2019-12-30 00:10:00     TRUE
#> 2           1 149.3067 -37.77333      2 2019-12-30 02:00:00    FALSE
#> 3           1 149.2800 -37.80000      3 2019-12-30 03:00:00    FALSE
#> 4           1 149.3200 -37.78000      4 2019-12-30 04:00:00    FALSE
#> 5           1 149.3100 -37.77000      5 2019-12-30 05:00:00    FALSE
#> 6           1 149.3200 -37.78000      7 2019-12-30 06:20:00    FALSE
#> 7           1 149.3100 -37.77000     10 2019-12-30 09:50:00    FALSE
#> 8           1 149.2900 -37.77000     24 2019-12-30 23:40:00    FALSE
#> 9           1 149.2867 -37.75333     42 2019-12-31 17:50:00    FALSE
#> 10          1 149.2850 -37.75499     43 2019-12-31 19:00:00    FALSE
#> 11          1 149.2900 -37.74999     44 2019-12-31 20:00:00    FALSE
#> 12          1 149.2800 -37.74999     45 2019-12-31 20:10:00    FALSE
#> 13          1 149.2800 -37.74999     46 2019-12-31 21:40:00    FALSE
#> 14          1 149.2800 -37.74999     47 2019-12-31 22:50:00    FALSE
#> 15          1 149.2800 -37.74999     48 2019-12-31 23:40:00    FALSE
#> 16          1 149.2800 -37.74666     49 2020-01-01 01:00:00    FALSE
#> 17          1 149.2800 -37.75999     50 2020-01-01 02:00:00    FALSE
#> 18          1 149.2800 -37.74000     51 2020-01-01 02:50:00    FALSE
#> 19          1 149.2800 -37.74999     52 2020-01-01 03:10:00    FALSE
#> 20          1 149.2800 -37.75999     53 2020-01-01 04:50:00    FALSE
#> 21          1 149.2800 -37.74999     54 2020-01-01 05:10:00    FALSE
#> 22          1 149.2733 -37.76666     65 2020-01-01 16:50:00    FALSE
#> 23          1 149.2733 -37.76333     66 2020-01-01 18:00:00    FALSE
#> 24          1 149.2800 -37.74000     67 2020-01-01 18:50:00    FALSE
#> 25          1 149.2800 -37.74999     68 2020-01-01 20:00:00    FALSE
#> 26          1 149.2600 -37.75999     69 2020-01-01 20:40:00    FALSE
#> 27          1 149.2800 -37.74000     70 2020-01-01 21:50:00    FALSE
#> 28          1 149.2800 -37.74999     71 2020-01-01 23:00:00    FALSE
#> 29          1 149.2800 -37.74999     73 2020-01-02 00:40:00    FALSE
#> 30          1 149.2800 -37.74999     74 2020-01-02 01:50:00    FALSE
#> 31          1 149.2800 -37.74999     75 2020-01-02 02:30:00    FALSE
#> 32          1 149.2800 -37.74000     76 2020-01-02 03:30:00    FALSE
#> 33          1 149.2800 -37.74000     77 2020-01-02 04:20:00    FALSE
#> 34          1 149.2800 -37.74000     78 2020-01-02 05:30:00    FALSE
#> 35          1 149.2700 -37.74999     87 2020-01-02 14:20:00    FALSE
#> 36          1 149.2714 -37.76285     88 2020-01-02 16:00:00    FALSE
#> 37          1 149.2640 -37.73600     89 2020-01-02 17:00:00    FALSE
#> 38          1 149.2560 -37.75199     90 2020-01-02 18:00:00    FALSE
#> 39          1 149.2500 -37.76499     91 2020-01-02 19:00:00    FALSE
#> 40          1 149.2600 -37.74444     92 2020-01-02 20:00:00    FALSE
#> 41          1 149.2650 -37.74000     93 2020-01-02 21:00:00    FALSE
#> 42          1 149.2600 -37.75999     94 2020-01-02 21:30:00    FALSE
#> 43          1 149.2640 -37.74000     95 2020-01-02 23:00:00    FALSE
#> 44          1 149.2600 -37.74000     97 2020-01-03 01:00:00    FALSE
#> 45          1 149.2733 -37.75333     98 2020-01-03 01:30:00    FALSE
#> 46          1 149.2800 -37.74000     99 2020-01-03 03:00:00    FALSE
#> 47          1 149.2400 -37.75999    113 2020-01-03 16:40:00    FALSE
#> 48          1 149.2300 -37.75999    115 2020-01-03 19:00:00    FALSE
#> 49          1 149.2600 -37.75999    116 2020-01-03 19:50:00    FALSE
#> 50          1 149.2400 -37.75999    117 2020-01-03 20:20:00    FALSE

get_fire_mov(result, cluster = 2, step = 3, method = "median")

#>    membership    lon    lat timeID             obsTime ignition
#> 1           2 146.72 -36.84    229 2020-01-08 12:40:00     TRUE
#> 2           2 146.72 -36.84    232 2020-01-08 15:30:00    FALSE
#> 3           2 146.72 -36.84    235 2020-01-08 19:00:00    FALSE
#> 4           2 146.72 -36.82    238 2020-01-08 22:00:00    FALSE
#> 5           2 146.72 -36.82    241 2020-01-09 01:00:00    FALSE
#> 6           2 146.72 -36.82    244 2020-01-09 03:40:00    FALSE
#> 7           2 146.72 -36.84    254 2020-01-09 14:00:00    FALSE
#> 8           2 146.72 -36.80    258 2020-01-09 18:00:00    FALSE
#> 9           2 146.72 -36.82    261 2020-01-09 20:40:00    FALSE
#> 10          2 146.74 -36.83    282 2020-01-10 17:10:00    FALSE
#> 11          2 146.76 -36.86    304 2020-01-11 16:00:00    FALSE
#> 12          2 146.76 -36.86    307 2020-01-11 18:50:00    FALSE
#> 13          2 146.76 -36.87    310 2020-01-11 22:00:00    FALSE
#> 14          2 146.78 -36.90    327 2020-01-12 15:00:00    FALSE
#> 15          2 146.78 -36.90    330 2020-01-12 17:30:00    FALSE
#> 16          2 146.78 -36.90    334 2020-01-12 21:10:00    FALSE
#> 17          2 146.78 -36.90    337 2020-01-13 00:10:00    FALSE
#> 18          2 146.80 -36.88    359 2020-01-13 22:30:00    FALSE
#> 19          2 146.82 -36.86    377 2020-01-14 17:00:00    FALSE